backend/requirements.txt
google-generativeai
vellum-ai
python-dotenv
fastapi
uvicorn
pydantic
requests
python-multipart
aiofiles
flask
mss
opencv-python
numpy
pytesseract
yt_dlp
twelvelabs
pygame
PyPDF2
